requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
appdirs==1.4.4
asttokens==2.2.1
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.2.0
backcall==0.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
comm==0.1.2
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
dill==0.3.6
docker-pycreds==0.4.0
entrypoints==0.4
executing==1.2.0
filelock==3.9.0
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
gitdb==4.0.10
GitPython==3.1.30
htmlmin==0.1.12
idna==3.4
ImageHash==4.3.1
importlib-metadata==6.0.0
imutils==0.5.4
ipykernel==6.16.2
ipython==7.34.0
ipywidgets==8.0.4
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jupyter_client==7.4.9
jupyter_core==4.12.0
jupyterlab-widgets==3.0.5
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.5.3
matplotlib-inline==0.1.6
multidict==6.0.4
multimethod==1.9.1
multiprocess==0.70.14
nest-asyncio==1.5.6
networkx==2.6.3
numpy==1.21.6
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
opencv-python==4.7.0.68
packaging==23.0
pandas==1.3.5
pandas-profiling==3.6.6
parso==0.8.3
pathtools==0.1.2
patsy==0.5.3
pexpect==4.8.0
phik==0.12.3
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==2.6.2
prompt-toolkit==3.0.36
protobuf==4.21.12
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==11.0.0
pydantic==1.10.4
Pygments==2.14.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==25.0.0
regex==2022.10.31
requests==2.28.2
responses==0.18.0
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.12.2
sentry-sdk==1.15.0
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
stack-data==0.6.2
statsmodels==0.13.5
tangled-up-in-unicode==0.2.0
threadpoolctl==3.1.0
tokenizers==0.13.2
torch==1.13.1
torchvision==0.14.1
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
typeguard==2.13.3
typing_extensions==4.5.0
urllib3==1.26.14
visions==0.7.5
wandb==0.13.10
wcwidth==0.2.6
widgetsnbextension==4.0.5
xxhash==3.2.0
yarl==1.8.2
ydata-profiling==4.0.0
zipp==3.13.0
